requirements.txt
opendataloader_pdf
pyyaml
pytest
